requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.17.0
supabase>=2.0.0
python-dotenv>=1.0.0
whop-sdk>=0.1.0
